+++ BraintreeCore/src/main/java/com/braintreepayments/api/AnalyticsDatabase.kt
@@ -8,11 +8,13 @@ import androidx.room.RoomDatabase
 
 // Ref: https://developer.android.com/training/data-storage/room/migrating-db-versions
 @Database(
-        version = 3,
+        version = 5,
         entities = [AnalyticsEvent::class],
         autoMigrations = [
-            AutoMigration(from = 1, to = 3),
-            AutoMigration(from = 2, to = 3)
+            AutoMigration(from = 1, to = 2),
+            AutoMigration(from = 2, to = 3),
+            AutoMigration(from = 3, to = 4),
+            AutoMigration(from = 4, to = 5)
         ]
 )
 internal abstract class AnalyticsDatabase : RoomDatabase() {